Remove font and move styling to stylesheet
The fonts were being cut off in certain areas and so the decision was
made to just remove the custom font since it was not critical to the
project. All styling for the elements that was in the JavaScript was
moved to the stylesheet because the original error was found as to why
the JavaScript workaround was used in the first place. This has helped
clean up the script file.